Move Player Layout views to be together
And remove the dependency on a presenter
