+++ src/Views/GamePlayViews/PlayerLayout/PlayerLayout.tsx
@@ -1,31 +1,31 @@
 import { Component, ReactElement } from 'react'
-import GamePresenter from '../../InterfaceAdapters/GamePresenter/GamePresenter'
-import PlayerLayoutData from '../../InterfaceAdapters/GamePresenter/PlayerLayoutData'
+import PlayerLayoutData from '../../../InterfaceAdapters/GamePresenter/PlayerLayoutData'
+import PlayerLayoutDisplayData from './PlayerLayoutDisplayData'
 import PlayerTurnBox from './PlayerTurnBox'
 import './PlayerLayout.css'
 
 type Props = {
-  presenter: GamePresenter
+  allPlayerData: PlayerLayoutDisplayData
 }
 
 class PlayerLayout extends Component<Props> {
   render() {
-    const { presenter } = this.props
+    const { allPlayerData } = this.props
     return (
       <div id='player-layout'>
         <div></div>
         <div></div>
-        {this.renderPlayerTurnBox(presenter.getDataForPlayerAcross())}
+        {this.renderPlayerTurnBox(allPlayerData.dataForPlayerAcross)}
         <div></div>
         <div></div>
         <div></div>
-        {this.renderPlayerTurnBox(presenter.getDataForPlayerToLeft())}
+        {this.renderPlayerTurnBox(allPlayerData.dataForPlayerToLeft)}
         <div></div>
-        {this.renderPlayerTurnBox(presenter.getDataForPlayerToRight())}
+        {this.renderPlayerTurnBox(allPlayerData.dataForPlayerToRight)}
         <div></div>
         <div></div>
         <div></div>
-        {this.renderPlayerTurnBox(presenter.getDataForLocalPlayer())}
+        {this.renderPlayerTurnBox(allPlayerData.dataForLocalPlayer)}
         <div></div>
         <div></div>
       </div>
