+++ src/Views/GamePlayViews/PlayerLayout/PlayerLayoutDisplayData.ts
@@ -1,0 +1,10 @@
+import PlayerLayoutData from '../../../InterfaceAdapters/GamePresenter/PlayerLayoutData'
+
+interface PlayerLayoutDisplayData {
+  dataForLocalPlayer: PlayerLayoutData
+  dataForPlayerAcross: PlayerLayoutData
+  dataForPlayerToLeft: PlayerLayoutData
+  dataForPlayerToRight: PlayerLayoutData
+}
+
+export default PlayerLayoutDisplayData
